Added new version 2.1. See changelog for info
git-svn-id: http://plugins.svn.wordpress.org/two-factor-auth/tags/2.1@703345 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ two-factor-login.php
@@ -2,12 +2,19 @@
 /*
 Plugin Name: Two Factor Auth
 Plugin URI: http://oskarhane.com/plugin-two-factor-auth-for-wordpress
-Description: Add extra security to your WordPress login with this two factor auth. Users will be prompted with a page to enter a code that was emailed to them.
+Description: Secure your WordPress login with this two factor auth. Users will be prompted with a page to enter a one time code that was emailed to them.
 Author: Oskar Hane
 Author URI: http://oskarhane.com
-Version: 1.0
+Version: 2.1
 License: GPLv2 or later
 */
+//error_reporting(E_ALL);
+//ini_set("display_errors", true);
+
+function showAdminSettingsPage()
+{
+	include 'admin_settings.php';
+}
 
 
 function breakAuth($log)
@@ -21,10 +28,27 @@ function breakAuth($log)
 	
 	if(!$params['two_factor_code_submitted'])
 		loadTwoFactorForm($params);
-	else
-		checkTwoFactorCode($params);
 }
 add_action('wp_authenticate', 'breakAuth');
+
+
+
+function tfaVerifyCodeAndUser($user, $username, $password)
+{
+	//If already failed, we don't bother to check the code
+	if(is_wp_error($user))
+		return $user;
+		
+	$params = $_POST;
+	$code_ok = checkTwoFactorCode($params);
+	
+	if(!$code_ok)
+		return new WP_Error('authentication_failed', __('<strong>ERROR</strong>: The Two Factor Code you entered was incorrect.'));
+
+	return wp_authenticate_username_password(null, $username, $password);
+}
+add_filter('authenticate', 'tfaVerifyCodeAndUser', 99999999999, 3);//We want to be the last filter that runs.
+
 
 
 function loadTwoFactorForm($params)
@@ -35,18 +59,43 @@ function loadTwoFactorForm($params)
 		
 	$code = generateTwoFactorCode();
 	
+	//So we show full form for users that dont exist
+	$is_activated_for_user = true;
+
 	//Render form anyway so we don't reveal if the username exists or not
 	if($user)
 	{
-		delete_user_meta($user->ID, 'two_factor_login_code');
-		add_user_meta($user->ID, 'two_factor_login_code', $code);
-		sendTwoFactorEmail($user->user_email, $code);
+		$is_activated_for_user = tfaIsActivatedForUser($user->ID);
+		
+		if($is_activated_for_user)
+		{
+			delete_user_meta($user->ID, 'two_factor_login_code');
+			add_user_meta($user->ID, 'two_factor_login_code', $code);
+			sendTwoFactorEmail($user->user_email, $code);
+		}
 	}
 	
 	include 'form.php';
 	
 	
 	exit;
+}
+
+
+function tfaIsActivatedForUser($user_id)
+{
+	$user = new WP_User($user_id);
+
+	foreach($user->roles as $role)
+	{
+		$db_val = get_option('tfa_'.$role);
+		$db_val = $db_val === false || $db_val ? 1 : 0; //Nothing saved or > 0 returns 1;
+		
+		if($db_val)
+			return true;
+	}
+	
+	return false;
 }
 
 
@@ -73,17 +122,26 @@ function checkTwoFactorCode($params)
 	$query = $wpdb->prepare("SELECT ID from ".$wpdb->users." WHERE user_login=%s", $params['log']);
 	$user_ID = $wpdb->get_var($query);
 	
+	if(!$user_ID)
+		return true;
+	
 	$code = get_user_meta($user_ID, 'two_factor_login_code', true);
 	
 	//Remove code after one guess
 	delete_user_meta($user_ID, 'two_factor_login_code');
 	
+	if(!tfaIsActivatedForUser($user_ID))
+		return true;
+	
 	if(!$code)
-		sendBackToLogin();
+		return false;
 
 	if($code != strtoupper($params['two_factor_code']))
-		sendBackToLogin();
+		return false;
+		
+	return true;
 }
+
 
 
 function sendTwoFactorEmail($email, $code)
@@ -92,21 +150,71 @@ function sendTwoFactorEmail($email, $code)
 }
 
 
-function sendBackToLogin()
+function generateTwoFactorCode()
 {
-	header('Location: '.$_SERVER['REQUEST_URI']);
-	exit;
+	$chars = '23456789QWERTYUPASDFGHJKLZXCVBNM';
+	$chars = str_split($chars);
+	shuffle($chars);
+	$code = implode('', array_splice($chars, 0, rand(5,6)));
+	
+	return $code;
 }
 
 
-function generateTwoFactorCode($len = 5)
+
+/**
+* For Admin menu and settings
+*/
+function registerTwoFactorAuthSettings()
 {
-	$chars = '123456789QWERTYUIPASDFGHJKLZXCVBNM';
-	$chars = str_split($chars);
-	shuffle($chars);
-	$code = implode('', array_splice($chars, 0, $len));
+	global $wp_roles;
+	if (!isset($wp_roles))
+		$wp_roles = new WP_Roles();
 	
-	return $code;
+	foreach($wp_roles->role_names as $id => $name)
+	{
+		register_setting('tfa_user_roles_group', 'tfa_'.$id);
+	}
+	
+}
+
+function tfaListUserRolesCheckboxes()
+{
+	global $wp_roles;
+	if (!isset($wp_roles))
+		$wp_roles = new WP_Roles();
+	
+	foreach($wp_roles->role_names as $id => $name)
+	{	
+		$setting = get_option('tfa_'.$id);
+		$setting = $setting === false || $setting ? 1 : 0;
+		
+		print '<input type="checkbox" name="tfa_'.$id.'" value="1" '.($setting ? 'checked="checked"' :'').'> '.$name."<br>\n";
+	}
+	
+}
+
+function addTwoFactorAuthAdminMenu()
+{
+	add_action( 'admin_init', 'registerTwoFactorAuthSettings' );
+	add_options_page('Two Factor Auth', 'Two Factor Auth', 'manage_options', 'two-factor-auth', 'showAdminSettingsPage');
+}
+
+function addPluginSettingsLink($links)
+{
+	$link = '<a href="options-general.php?page=two-factor-auth">'.__('Settings').'</a>';
+	array_unshift($links, $link);
+	return $links;
+}
+
+if(is_admin())
+{
+	//Add to Settings menu
+	add_action('admin_menu', 'addTwoFactorAuthAdminMenu');
+	
+	//Add settings link in plugin list
+	$plugin = plugin_basename(__FILE__); 
+	add_filter("plugin_action_links_".$plugin, 'addPluginSettingsLink' );
 }
 
 ?>
